split query and index into submodules
